fix: Removed property change in all combined filters, reduced floating point precision in SIN and COS.
Piklib seems to be doing trigonometric calculations in float, not double. I'm testing the change, but it's probably not necessary.

+++ core/src/pl/genschu/bloomooemulator/interpreter/variable/types/DoubleVariable.java
@@ -129,7 +129,7 @@ public class DoubleVariable extends Variable {
 			public Variable execute(List<Object> arguments) {
 				double angle = ArgumentsHelper.getDouble(arguments.get(0));
 
-				double cosRadians = Math.cos(Math.toRadians(angle));
+				double cosRadians = Math.cos((float) Math.toRadians(angle));
 
 				set(cosRadians);
 				return DoubleVariable.this;
@@ -332,7 +332,7 @@ public class DoubleVariable extends Variable {
 			public Variable execute(List<Object> arguments) {
 				double angle = ArgumentsHelper.getDouble(arguments.get(0));
 
-				double sinRadians = Math.sin(Math.toRadians(angle));
+				double sinRadians = Math.sin((float) Math.toRadians(angle));
 
 				set(sinRadians);
 				return DoubleVariable.this;
